+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## master
+
+- Upgrade `graphiql-explorer` to latest version.
+- Add "Edit" operation as code action.
+
 ## 1.2.0
 
 - Upgrade `graphiql-explorer` to latest version.